Update tests in the storage module
- Remove old tests.rs file
- Introduce a new tests module
- Move mock and all tests into the tests module
