+++ runtime-modules/storage/src/tests/mod.rs
@@ -1,0 +1,6 @@
+#![cfg(test)]
+
+mod data_directory;
+mod data_object_storage_registry;
+mod data_object_type_registry;
+mod mock;
